Move kyber testing-convenience functions, to break import cycle
